# Conflicts:
#	arena-api/src/main/java/io/github/zap/arenaapi/game/arena/ManagedPlayer.java
#	arena-api/src/main/java/io/github/zap/arenaapi/game/arena/ManagingArena.java
#	zombies/src/main/java/io/github/zap/zombies/game/ZombiesArena.java

+++ zombies/src/main/java/io/github/zap/zombies/game/ZombiesArena.java
@@ -3,6 +3,7 @@ package io.github.zap.zombies.game;
 import com.comphenix.protocol.PacketType;
 import com.comphenix.protocol.events.PacketContainer;
 import io.github.zap.arenaapi.Property;
+import io.github.zap.arenaapi.ResourceManager;
 import io.github.zap.arenaapi.event.Event;
 import io.github.zap.arenaapi.event.EventHandler;
 import io.github.zap.arenaapi.event.FilteredEvent;
@@ -53,6 +54,7 @@ import org.bukkit.entity.Entity;
 import org.bukkit.entity.ItemFrame;
 import org.bukkit.entity.Mob;
 import org.bukkit.entity.Player;
+import org.bukkit.event.HandlerList;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.Action;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
@@ -63,7 +65,6 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.event.player.*;
 import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.scheduler.BukkitScheduler;
 import org.bukkit.util.Vector;
 import org.jetbrains.annotations.NotNull;
 
@@ -483,6 +484,8 @@ public class ZombiesArena extends ManagingArena<ZombiesArena, ZombiesPlayer> imp
 
         getMap().getPowerUpSpawnRules()
                 .forEach(x -> powerUpSpawnRules.add(ImmutablePair.of(getPowerUpManager().createSpawnRule(x.left, x.right, this), x.right)));
+
+        Bukkit.getServer().getPluginManager().registerEvents(this, Zombies.getInstance());
     }
 
     private void registerMythicMobsEvents() {
@@ -514,10 +517,12 @@ public class ZombiesArena extends ManagingArena<ZombiesArena, ZombiesPlayer> imp
     }
 
     private void registerDisposables() {
-        getResourceManager().addDisposable(mythicMobDeathEvent);
-        getResourceManager().addDisposable(mythicMobDespawnEvent);
-        getResourceManager().addDisposable(gameScoreboard);
-        getResourceManager().addDisposable(powerUpBossBar);
+        ResourceManager resourceManager = getResourceManager();
+
+        resourceManager.addDisposable(mythicMobDeathEvent);
+        resourceManager.addDisposable(mythicMobDespawnEvent);
+        resourceManager.addDisposable(gameScoreboard);
+        resourceManager.addDisposable(powerUpBossBar);
     }
 
     @Override
@@ -530,13 +535,10 @@ public class ZombiesArena extends ManagingArena<ZombiesArena, ZombiesPlayer> imp
         super.dispose(); //dispose of superclass-specific resources
 
         //cleanup mappings and remove arena from manager
-        for (Shop<?> shop : shopMap.get(ShopType.TEAM_MACHINE)) {
-            TeamMachine teamMachine = (TeamMachine) shop;
-            Property.removeMappingsFor(teamMachine);
-        }
-
         Property.removeMappingsFor(this);
         manager.unloadArena(getArena());
+
+        HandlerList.unregisterAll(this);
     }
 
     @Override
@@ -639,6 +641,15 @@ public class ZombiesArena extends ManagingArena<ZombiesArena, ZombiesPlayer> imp
         }
     }
 
+    @org.bukkit.event.EventHandler
+    private void onNonManagedEntityDamage(EntityDamageEvent event) {
+        Entity entity = event.getEntity();
+
+        if (entity.getWorld().equals(world) && entity instanceof ItemFrame) {
+            event.setCancelled(true);
+        }
+    }
+
     private void onEntityDamageByEntity(ProxyArgs<EntityDamageByEntityEvent> args) {
         EntityDamageByEntityEvent event = args.getEvent();
         Entity entity = event.getEntity(), damager = event.getDamager();
@@ -668,46 +679,49 @@ public class ZombiesArena extends ManagingArena<ZombiesArena, ZombiesPlayer> imp
     private void onPlayerDeath(ProxyArgs<PlayerDeathEvent> args) {
         args.getEvent().setCancelled(true); //cancel death event
 
-        if(state == ZombiesArenaState.STARTED) {
+        if (state == ZombiesArenaState.STARTED) {
             ZombiesPlayer knocked = args.getManagedPlayer();
-            knocked.knock();
 
-            for(ZombiesPlayer player : getPlayerMap().values()) {
-                if(player.isAlive()) {
-                    Player knockedBukkitPlayer = knocked.getPlayer();
-                    RoomData knockedRoom = map.roomAt(knockedBukkitPlayer.getLocation().toVector());
-                    String message = knockedRoom == null ? "an unknown room" : knockedRoom.getRoomDisplayName();
+            if (knocked != null) {
+                knocked.knock();
 
-                    //display death message only if necessary
-                    for(ZombiesPlayer otherPlayer : getPlayerMap().values()) {
-                        if(otherPlayer != knocked) {
-                            otherPlayer.getPlayer().showTitle(Title.title(Component.text(knockedBukkitPlayer.getName())
-                                    .color(TextColor.color(255, 255, 0)), Component.text("was knocked down in " + message)
-                                    .color(TextColor.color(61, 61, 61)), Title.Times.of(Duration.ofSeconds(1),
-                                    Duration.ofSeconds(3), Duration.ofSeconds(1))));
+                for (ZombiesPlayer player : getPlayerMap().values()) {
+                    if (player.isAlive()) {
+                        Player knockedBukkitPlayer = knocked.getPlayer();
+                        RoomData knockedRoom = map.roomAt(knockedBukkitPlayer.getLocation().toVector());
+                        String message = knockedRoom == null ? "an unknown room" : knockedRoom.getRoomDisplayName();
+
+                        //display death message only if necessary
+                        for (ZombiesPlayer otherPlayer : getPlayerMap().values()) {
+                            if (otherPlayer != knocked) {
+                                otherPlayer.getPlayer().showTitle(Title.title(Component.text(knockedBukkitPlayer.getName())
+                                        .color(TextColor.color(255, 255, 0)), Component.text("was knocked down in " + message)
+                                        .color(TextColor.color(61, 61, 61)), Title.Times.of(Duration.ofSeconds(1),
+                                        Duration.ofSeconds(3), Duration.ofSeconds(1))));
 
 
-                            otherPlayer.getPlayer().playSound(Sound.sound(
-                                    Key.key("minecraft:entity.ender_dragon.growl"),
-                                    Sound.Source.MASTER,
-                                    1.0F,
-                                    0.5F
-                            ));
+                                otherPlayer.getPlayer().playSound(Sound.sound(
+                                        Key.key("minecraft:entity.ender_dragon.growl"),
+                                        Sound.Source.MASTER,
+                                        1.0F,
+                                        0.5F
+                                ));
+                            }
                         }
+
+                        return; //return if there are any players still alive
                     }
-
-                    return; //return if there are any players still alive
                 }
-            }
 
-            doLoss(); //there are no players alive, so end the game
+                doLoss(); //there are no players alive, so end the game
 
-            // Bit hacky way to make sure corpses are registered to a team before their holograms are destroyed
-            for(ZombiesPlayer player : getPlayerMap().values()) {
-                player.kill();
-                Corpse corpse = player.getCorpse();
-                if (corpse != null) {
-                    corpse.terminate();
+                // Bit hacky way to make sure corpses are registered to a team before their holograms are destroyed
+                for (ZombiesPlayer player : getPlayerMap().values()) {
+                    player.kill();
+                    Corpse corpse = player.getCorpse();
+                    if (corpse != null) {
+                        corpse.terminate();
+                    }
                 }
             }
         }
@@ -734,11 +748,11 @@ public class ZombiesArena extends ManagingArena<ZombiesArena, ZombiesPlayer> imp
         event.setCancelled(true);
     }
 
-    private void onPlayerInteractEntity(ProxyArgs<PlayerInteractEntityEvent> args) {
-        PlayerInteractEntityEvent event = args.getEvent();
-        Entity clickedEntity = event.getRightClicked();
+    @org.bukkit.event.EventHandler
+    private void onPlayerInteractEntity(PlayerInteractEntityEvent event) {
+        Entity entity = event.getRightClicked();
 
-        if (clickedEntity instanceof ItemFrame) {
+        if (entity.getWorld().equals(world) && entity instanceof ItemFrame) {
             event.setCancelled(true);
         }
     }
@@ -1048,6 +1062,11 @@ public class ZombiesArena extends ManagingArena<ZombiesArena, ZombiesPlayer> imp
             shop.display();
         }
         getShopEvent(ShopType.DOOR);
+
+        for (Shop<?> shop : shopMap.get(ShopType.TEAM_MACHINE)) {
+            TeamMachine teamMachine = (TeamMachine) shop;
+            getResourceManager().addDisposable(teamMachine);
+        }
 
         Event<ShopEventArgs> chestEvent = shopEvents.get(ShopType.LUCKY_CHEST);
         if (chestEvent != null) {
